Update README.md
In this Project, I cleaned and analyzed Sales Data for a Bike Company and built a beautiful user friendly Dashboard that shows insights of how customers purchased the bikes. 

After importing the Dataset into Excel, I did a Data overview to remove duplicates, used the "find and replace" method to properly show values in multiple cells, formatted the column for income with the right currency and used the IF function (Nested IF) to create another column for Age bracket to make visualization appealing. 

Created Pivot tables and Charts to visualized the cleaned data and built the Dashboard with header, proper alignments, design and finally slicers that connects the entire charts to make it interactive for users.